🏷️ feat: Added AirInterfaceUserRate enum
This commit introduces the `AirInterfaceUserRate` enum, which is used to represent the air interface user rate in Ericsson telecommunication data. The enum maps integer values to their corresponding string representations, as defined in the ASN.1 specification. This enhancement improves the ability to decode and interpret air interface user rate information.