🏷️ feat: Added NumberOfMeterPulses datatype
This commit introduces the `NumberOfMeterPulses` datatype, which is used to represent the number of meter pulses in Ericsson telecommunication data. The datatype is defined as an OCTET STRING of size 3, coded as an unsigned integer. This enhancement improves the ability to decode and interpret meter pulse information. Additionally, this commit updates the base classes for `CallIDNumber`, `RecordSequenceNumber`, and `TypeOfCallingSubscriber` to inherit from `primitives.UnsignedInt` and `NumberOfMeterPulses` to inherit from `primitives.DigitString`.