🏷️ feat: Added INMarkingOfMS datatype
This commit introduces the `INMarkingOfMS` datatype, which is used to represent the Intelligent Network (IN) marking of the Mobile Station (MS) in Ericsson telecommunication data. The datatype is defined as an ENUMERATED type with thirteen possible values, each representing a different IN service or CAMEL service. This enhancement improves the ability to decode and interpret IN service information.