🏷️ feat: Added UserClass datatype
This commit introduces the `UserClass` datatype, a `ByteEnum` used to represent the type of phone used on the originating side of a call in WCDMA Japan networks. It is mapped from the Additional Fixed User Information in the Additional User Category (AUC) parameter in ISUP message IAM.

The datatype includes an ASN.1 formal description and a set of enumerated values representing different user classes, such as "Train Payphone" and "Pink (non-NTT Payphone)". The `value` property returns the spare bits and the additional information, and a `__str__` method is added for a user-friendly representation.