feat: Added CarrierInfo datatype
This commit introduces the `CarrierInfo` datatype, an implementation of an OCTET STRING with a size of 2-3 octets. It parses the octets to extract the Carrier Identification Code and the Entry/Exit POI Hierarchy, providing a structured way to access these values. This datatype is used in Ericsson call data records.